selenium-webdriver: move logging to separate file
