+++ types/selenium-webdriver/test/index.ts
@@ -522,55 +522,6 @@ function TestWebElementPromise() {
     elementPromise.then((element: webdriver.WebElement) => 'foo', (error: any) => {}).then((result: string) => {});
 }
 
-function TestLogging() {
-    let preferences: webdriver.logging.Preferences = new webdriver.logging.Preferences();
-    preferences.setLevel(webdriver.logging.Type.BROWSER, webdriver.logging.Level.ALL);
-    let prefs: any = preferences.toJSON();
-
-    let level: webdriver.logging.Level = webdriver.logging.getLevel('OFF');
-    level = webdriver.logging.getLevel(1);
-
-    level = webdriver.logging.Level.ALL;
-    level = webdriver.logging.Level.DEBUG;
-    level = webdriver.logging.Level.INFO;
-    level = webdriver.logging.Level.OFF;
-    level = webdriver.logging.Level.SEVERE;
-    level = webdriver.logging.Level.WARNING;
-
-    let name: string = level.name;
-    let value: number = level.value;
-
-    let type: string;
-    type = webdriver.logging.Type.BROWSER;
-    type = webdriver.logging.Type.CLIENT;
-    type = webdriver.logging.Type.DRIVER;
-    type = webdriver.logging.Type.PERFORMANCE;
-    type = webdriver.logging.Type.SERVER;
-
-    let logger: webdriver.logging.Logger = webdriver.logging.getLogger();
-    webdriver.logging.addConsoleHandler();
-    webdriver.logging.addConsoleHandler(logger);
-    webdriver.logging.removeConsoleHandler();
-    webdriver.logging.removeConsoleHandler(logger);
-}
-
-function TestLoggingEntry() {
-    let entry: webdriver.logging.Entry;
-
-    entry = new webdriver.logging.Entry(webdriver.logging.Level.ALL, 'ABC');
-    entry = new webdriver.logging.Entry('ALL', 'ABC');
-    entry = new webdriver.logging.Entry(webdriver.logging.Level.ALL, 'ABC', 123);
-    entry = new webdriver.logging.Entry('ALL', 'ABC', 123);
-    entry = new webdriver.logging.Entry(webdriver.logging.Level.ALL, 'ABC', 123, webdriver.logging.Type.BROWSER);
-    entry = new webdriver.logging.Entry('ALL', 'ABC', 123, webdriver.logging.Type.BROWSER);
-
-    let entryObj: any = entry.toJSON();
-
-    let message: string = entry.message;
-    let timestamp: number = entry.timestamp;
-    let type: string = entry.type;
-}
-
 declare let stringPromise: Promise<string>;
 
 function TestUntilModule() {
